rename PAE Jiffy component parse out teh instructions to seperate file
